+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - renderable: added a basic `NineSliceSprite` renderable object
+- bounds: faster condition assessment for the `overlaps` method
 
 ## [10.1.1] (melonJS 2) - _2021-11-12_